Refactor activity logging
