--- lib/trento/infrastructure/activity_log/logger/adapter/noop_logger.ex
@@ -1,10 +1,0 @@
-defmodule Trento.Infrastructure.ActivityLog.Logger.NoopLogger do
-  @moduledoc """
-  Noop Activity Logger Adapter
-  """
-
-  @behaviour Trento.ActivityLog.ActivityLogger
-
-  @impl true
-  def log_activity(_), do: :ok
-end
