+++ test/support/conn_case.ex
@@ -17,6 +17,8 @@ defmodule TrentoWeb.ConnCase do
 
   use ExUnit.CaseTemplate
 
+  use Trento.TaskCase
+
   alias Ecto.Adapters.SQL.Sandbox
 
   using do
@@ -35,17 +37,13 @@ defmodule TrentoWeb.ConnCase do
 
   setup tags do
     pid = Sandbox.start_owner!(Trento.Repo, shared: not tags[:async])
-    on_exit(fn -> Sandbox.stop_owner(pid) end)
 
-    stub_activity_logger()
+    on_exit(fn ->
+      wait_for_tasks_completion()
+
+      Sandbox.stop_owner(pid)
+    end)
 
     {:ok, conn: Phoenix.ConnTest.build_conn()}
   end
-
-  defp stub_activity_logger,
-    do:
-      Mox.stub_with(
-        Trento.ActivityLog.ActivityLogger.Mock,
-        Trento.Infrastructure.ActivityLog.Logger.NoopLogger
-      )
 end
